Fix price page
