+++ src/components/price-page/index.tsx
@@ -1,14 +1,46 @@
 import { useEffect } from "react";
 import { useGetPrices } from "../../hooks/useGetPrices";
+import { BaseCurrency, PriceInfo } from "../../schema";
 import * as Styles from "./styles";
 
 interface PricePageProps {
   currencies: string[];
-  baseCurrency: string;
+  baseCurrency: BaseCurrency;
   refreshDelay: number;
   onUpdate: () => void;
   onLoading: (isLoading: boolean) => void;
 }
+
+const currencyToSymbol = (currency: BaseCurrency) => {
+  switch (currency) {
+    case BaseCurrency.EURO:
+      return "€";
+    case BaseCurrency.USDOLLAR:
+      return "$";
+  }
+};
+
+const getPrice = (currency: BaseCurrency, info: PriceInfo) => {
+  switch (currency) {
+    case BaseCurrency.EURO:
+      return info.eur ?? 0;
+    case BaseCurrency.USDOLLAR:
+      return info.usd ?? 0;
+    default:
+      return 0;
+  }
+};
+
+const getChange = (currency: BaseCurrency, info: PriceInfo) => {
+  switch (currency) {
+    case BaseCurrency.EURO:
+      return info.eur_24h_change ?? 0;
+    case BaseCurrency.USDOLLAR:
+      return info.usd_24h_change ?? 0;
+    default:
+      return 0;
+  }
+};
 
 const PricePage = (props: PricePageProps) => {
   const { currencies, onUpdate, baseCurrency, refreshDelay, onLoading } = props;
@@ -31,23 +63,27 @@ const PricePage = (props: PricePageProps) => {
   return (
     <Styles.CoinGrid items={prices.length}>
       {prices &&
-        prices.map((price, index) => (
-          <Styles.CoinItem key={index}>
-            <div>
-              <h2>{price.symbol}</h2>
-              <Styles.CoinValue
-                perc={Number.parseInt(price.eur_24h_change.toFixed(2))}
-              >
-                <span className="price">
-                  {price.eur.toFixed(price.decimals)}€
-                </span>
-                <span className="percentage">
-                  {price.eur_24h_change.toFixed(2)}%
-                </span>
-              </Styles.CoinValue>
-            </div>
-          </Styles.CoinItem>
-        ))}
+        prices.map((price, index) => {
+          const priceVal = getPrice(baseCurrency, price);
+          const priceChange = getChange(baseCurrency, price);
+
+          return (
+            <Styles.CoinItem key={index}>
+              <div>
+                <h2>{price.symbol}</h2>
+                <Styles.CoinValue
+                  perc={Number.parseInt(priceChange.toFixed(2))}
+                >
+                  <span className="price">
+                    {priceVal.toFixed(price.decimals)}{" "}
+                    {currencyToSymbol(baseCurrency)}
+                  </span>
+                  <span className="percentage">{priceChange.toFixed(2)}%</span>
+                </Styles.CoinValue>
+              </div>
+            </Styles.CoinItem>
+          );
+        })}
     </Styles.CoinGrid>
   );
 };
